Added support for reliable UDP congestion control and packet resending -- no transcieve yet #78

+++ server/netty/src/com/riiablo/net/reliable/data/ReceivedPacketData.java
@@ -1,6 +1,6 @@
 package com.riiablo.net.reliable.data;
 
 public class ReceivedPacketData {
-  public long    time;
+  public float   time;
   public int     packetSize;
 }
