+++ server/netty/src/com/riiablo/net/reliable/data/SentPacketData.java
@@ -1,7 +1,7 @@
 package com.riiablo.net.reliable.data;
 
 public class SentPacketData {
-  public long    time;
+  public float   time;
   public boolean acked;
   public int     packetSize;
 }
